Create dedicated file for eval/rule and eval/event
