+++ pkg/security/secl/eval/event.go
@@ -1,0 +1,6 @@
+package eval
+
+type Event interface {
+	GetID() string
+	GetType() string
+}
