+++ pkg/security/secl/eval/rule.go
@@ -1,0 +1,7 @@
+package eval
+
+type Rule struct {
+	Name       string
+	Expression string
+	evaluator  *RuleEvaluator
+}
